Refactored entire structure of API and backend
