+++ schemas/chat.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class QuestionRequest(BaseModel):
+    question: str
